api is complete

+++ backend/controllers/goalController.js
@@ -1,4 +1,5 @@
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 const asyncHandler = require('express-async-handler');
 
 const getGoals = asyncHandler(async (req, res) => {
@@ -29,6 +30,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error('Goal not found!');
   }
 
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error('User no found!');
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized!');
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -42,6 +55,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found!');
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error('User no found!');
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized!');
   }
 
   await Goal.findByIdAndDelete(req.params.id);
